refactor(oracle): migrate dns scapy encode to dns plugin
